% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Est_clust_area2.R
\name{Est_clust_area2}
\alias{Est_clust_area2}
\title{Use clustered coefficients models in area-level models}
\usage{
Est_clust_area2(
  dom,
  y,
  x,
  index_d,
  vardir,
  lamvec,
  nu = 1,
  gam = 3,
  lam0 = 0.001,
  maxiter = 500,
  tol = 1e-04,
  seed = 1234,
  init_control = NULL,
  beta00 = NULL
)
}
\arguments{
\item{dom}{1:m, m is the number of domains}

\item{y}{direct estimator}

\item{x}{covaraites including intercept}

\item{index_d}{the index with the cluster structure}

\item{vardir}{the direct variance}

\item{lamvec}{the sequence of tuning parameters lambda}
}
\description{
Use clustered coefficients models in area-level models
}
\examples{
library(CLAreaModel)
data(simdata)
lamvec <- 10^seq(-1.05, -0.5, length.out=50)
res_clust <- Est_clust_area2(dom = 1:nrow(df_area),
                             y = df_area$y, x = cbind(1, df_area$x),
                             index_d = 1:2, vardir = df_area$vardir,
                             lamvec = lamvec)
estCL <- res_clust$area_estm ### area estimate
clusterm <- res_clust$cluster_estm ## estimated number

}
